moved roi_stats() to extract.py
